apps/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.42
psycopg2-binary==2.9.10
alembic==1.16.4
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
itsdangerous==2.2.0
python-multipart==0.0.9

# Redis para cache
redis==6.2.0

# Testing
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1

# Logging y utilidades
loguru==0.7.3
python-dotenv==1.1.1

# Rate limiting y seguridad
slowapi==0.1.9
limits==5.4.0

# HTTP requests
httpx==0.28.1
requests==2.32.4

# Timezone y fechas
pytz==2025.2

# Development tools
black==25.1.0
isort==6.0.1
flake8==7.3.0
mypy==1.17.1